Provision pdftk
The pdf-forms gem requires software to be installed on the VM. I
added a provisioner to the Vangrantfile. If you have an existing
Vagrant machine in the directory where you're working on the project,
and it's running, type `vagrant provision` on the host. If it's not
running, type `vagrant up --provision` to force provisioning on
startup. You only have to do this once.