Save work.
Really struggling with using accept_nested_attributes_for and the
upside-down relationship: when the user is creating the form, we
want to be able to update the existing user record.

There are many issues. The fact that the back-end data model is
more general than the UI was hard. I think I mostly got it sorted
out when I realized that some of the `fields_for` magic only works
when the items you're dealing with are actually attributes. And
any other approach to shoe-horning the attributes onto the right
phone number was challenging.

I read something about not being able to use the same parameter
cleanser for new and edit. I think that's for nested attributes
because they might not necessarily have the nested attributes.

I'm going to work on the Relationship project for a while. This is
too frustrating.