Allow save of empty RTP.
I changed the validations to be on the :printable context, so they
only fire when I ask them to. I generate a link only if the form
is printable. Otherwise they get a disabled useless button.

I had to change a lot of test cases, and I've put skip in a bunch.