Fix accordion on home page Issue #29
This is part of fixing Issue #29. Panels now have IDs based on the
child ID. The test case was very difficult. In the end, I abandoned
it because Bootstrap makes accordions work by setting the height
of the hidden stuff to 0, and I couldn't figure out a selector to
identify that case.

As a note, this happens because one shows a nice prototype, but
when you move from prototype to the real thing, there are a lot of
bits of hard-coding that you have to remember to remove.