Javascript for fiscal year.
The main struggle here was with the assets. I think I may have
precompiled the assets before, and that prevented tests from seeing
the changed assets, so I was trying to make something work that
wasn't getting to poltergeist. Development worked, which was the
confusing part for a long time. It works as it should in test as
long as you don't have any precompiled assets.