Make status report a table.
I tried this once and it looked ugly. Found stuff on the net that
said Bootstrap tables don't respect columns. Deleted everything and
tried some other stuff. As usual, CSS was a swamp. Looked at invoices
and said, "no, it's working for Phil.". Tried Bootply for tables
and they respected the columns. Tried again and it seems to work.

Go figure.