Preferences for panel state and fiscal year.
This implementation puts everything in User. It leads to rather
long calls to methods to get the values where they're needed.

The hard part was getting tests to work. It's very hard to sync
the test script with the Javascript execution. I finally put in
a sleep(5) because I was sick of the problem.