Fix test problems.
The PhantomJS tests we "randomly" failing. It turns out the
rollback from the transactional test fires after database_cleaner
has done its job, so the cleaning is rolled back and the database
is still full of old test data.

The fix is to put in the test class being used for the PhantomJS
tests, and derived from ActiveSupport::TestCase, the following:

self.use_transactional_tests = false

Put that right in the class, not in the setup method.