Update InvoiceAllocation on change to Cf0925
The implementation required adding a line to the Cf0925#create and
Cf0925#update actions, and also adding `dependent: :destroy` to
the association definition for `has_many :invoice_allocations`.

The Cf0925#create method had to be modified to build a Cf0925
on the funded_person's cf0925s association. This is probably a
general pattern that should be followed everywhere.